Higher priority for `symbol` than `pair` in response.
According to [the documentation](https://bitfinex.readme.io/v2/docs/ws-general#section-supported-pairs), one should prefix the pairs with a `t` for subscribing to the trading channel, and with an `f` for margin ones.

When the library uses `pair` with a higher priority, it means we can't subscribe to all variants on the same socket. Worse, we need to use `tBTCUSD` for subscribing, `BTCUSD` for accessing through `tickers('BTCUSD')`, and finally it's impossible to unsubscribe due to the mixup. You can try all these with the example in the README. This change fixes all of these problems.