Implemented single account queue
Removed the queue accessors from client.py and replaced it with a single 'account' queue accessor.

Turned the account queue in queue_processor.py from a dictionary into a normal single dimension Queue.
Removed the cumbersome translation dict in favor of having the end user figure out what the output is.

I am not sure if changing the identifier in _handle_account from 'auth' to 'account' will break something somewhere else.
In my own tests everything behaved as expected.

The self.update_timestamps() in run(), I couldn't quite figure out whether it was intended behaviour or not that in the previous version any account data update did not trigger an update of the timestamps.